requirements.txt
# Core Dependencies
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0

# Data Validation & Type Hints
# Note: xmlrpc.client is part of Python's standard library
typing-extensions>=4.8.0

# Logging
loguru>=0.7.2

# Testing (Optional but recommended)
pytest>=7.4.0
pytest-cov>=4.1.0

google-generativeai>=0.2.0

